Changed Data Type In The Class , Type Conversion Problem

+++ FashionableMe/FashionableMe/Models/Apparel.cs
@@ -10,7 +10,7 @@ namespace FashionableMe.Models
         public int ApparelID { get; set; }
         public string ApparelName { get; set; }
         public string BrandName { get; set; }
-        public double ApparelCost { get; set; }
+        public decimal ApparelCost { get; set; }
         public string Description { get; set; }
         public string ApparelImage { get; set; }
         public string ApparelCategory { get; set; }
@@ -18,7 +18,7 @@ namespace FashionableMe.Models
         public int ApparelRating { get; set; }
         public string ApparelSize { get; set; }
         public int QuantityPerSize { get; set; }
-        public double ApparelDiscount { get; set; }
+        public decimal ApparelDiscount { get; set; }
 
     }
 }
